Allow affiliations to carry project ownership information; transform profile images correctly
Summary:
This is sort of doing two things at once:

  - Add an "isOwner" flag to Project Affiliation to lay the groundwork for T237.
  - Rename the "QuickCreate" workflow to "Create" and funnel all creation
through it.
  - Reorganize the image transformation stuff and use it to correctly
crop/resize uploaded images.

Test Plan:
Created and edited projects and affailiations. Uploaded project, user, and
profile photos. Verified existing thumbnailing in Maniphest still works
properly.

Reviewed By: cadamo
Reviewers: cadamo, aran, jungejason, tuomaspelkonen
CC: aran, epriestley, cadamo
Differential Revision: 529

+++ src/applications/project/controller/editaffiliation/PhabricatorProjectAffiliationEditController.php
@@ -52,7 +52,12 @@ class PhabricatorProjectAffiliationEditController
 
       if (!strlen($affiliation->getRole())) {
         if ($affiliation->getID()) {
-          $affiliation->delete();
+          if ($affiliation->getIsOwner()) {
+            $affiliation->setRole('Owner');
+            $affiliation->save();
+          } else {
+            $affiliation->delete();
+          }
         }
       } else {
         $affiliation->save();
